+++ src/common/utils/config.js
@@ -74,10 +74,31 @@ const iexecGatewayMap = {
   },
 };
 
+const ensMap = {
+  1: {
+    // registry: '0x00000000000C2E074eC69A0dFb2997BA6C7d2e1e',
+    publicResolver: '0x4976fb03C32e5B8cfe2b6cCB31c09Ba78EBaBa41',
+  },
+  5: {
+    // registry: '0x00000000000C2E074eC69A0dFb2997BA6C7d2e1e',
+    publicResolver: '0x4B1488B7a6B320d2D721406204aBc3eeAa9AD329',
+  },
+  133: {
+    registry: '0x5f5B93fca68c9C79318d1F3868A354EE67D8c006',
+    publicResolver: '0x1347d8a1840A810B990d0B774A6b7Bb8A1bd62BB',
+  },
+  134: {
+    registry: '0x5f5B93fca68c9C79318d1F3868A354EE67D8c006',
+    publicResolver: '0x1347d8a1840A810B990d0B774A6b7Bb8A1bd62BB',
+  },
+};
+
 const getChainDefaults = ({ id, flavour }) => ({
   host: hostMap[id],
   hub: hubMap[flavour] && hubMap[flavour][id],
   sms: smsMap[flavour] && smsMap[flavour][id],
+  ensRegistry: ensMap[id] && ensMap[id].registry,
+  ensPublicResolver: ensMap[id] && ensMap[id].publicResolver,
   resultProxy: resultProxyMap[flavour] && resultProxyMap[flavour][id],
   ipfsGateway:
     (ipfsGatewayMap[flavour] &&